Refactor workflow evaluation components and improve code organization
• Move components to dedicated files
• Create new EvaluationEditView component
• Update router configuration
• Relocate useEvaluationForm composable
• Add new types for evaluations
