+++ packages/editor-ui/src/components/WorkflowEvaluation/types.ts
@@ -1,0 +1,13 @@
+export interface TestExecution {
+	lastRun: string | null;
+	errorRate: number | null;
+	metrics: Record<string, number>;
+}
+
+export interface TestListItem {
+	id: number;
+	name: string;
+	tagName: string;
+	testCases: number;
+	execution: TestExecution;
+}
